dev_requirements.txt
arxiv~=1.4.8
bertopic~=0.15.0
jupyter~=1.0.0
langchain~=0.0.263
open-clip-torch~=2.20.0
nltk~=3.8.1
numpy~=1.24.4
openai~=0.27.8
pandas~=2.0.3
plotly~=5.15.0
python-dotenv~=1.0.0
scikit-learn~=1.3.0
streamlit-plotly-events~=0.0.6
torch~=1.3.0
torchdiffeq~=0.2.3
torchsde~=0.2.5
einops~=0.6.1
transformers>=4.25.1
safetensors>=0.3.0